requirement.txt
speechrecognition
pillow
ipython
